+++ main.go
@@ -38,7 +38,7 @@ func main() {
 	for i := 0; i < 5; i++ {
 		h := &codec.Header{
 			ServiceMethod: "Foo.Sum",
-			Seq:           i,
+			Seq:           uint64(i),
 		}
 		_ = cc.Write(h, fmt.Sprintf("geerpc req %d", h.Seq))
 		_ = cc.ReadHeader(h)
